--- a/.travis.yml
+++ b/.travis.yml
@@ -18,7 +18,6 @@ install:
   - pip install -r dev-requirements.txt
   - python setup.py install
   - if [[ $TRAVIS_PYTHON_VERSION == "3.6" ]]; then pip install mypy; fi
-  - if [[ $TRAVIS_PYTHON_VERSION != "pypy" ]]; then pip install cython; fi
 
 script:
   - if [[ $TRAVIS_PYTHON_VERSION == "3.6" ]]; then mypy --ignore-missing-imports --follow-imports=silent --python-version=3.6 --warn-no-return --strict-optional toastedmarshmallow tests; fi