--- a/toastedmarshmallow/jit.py
+++ b/toastedmarshmallow/jit.py
@@ -13,9 +13,6 @@ from .compat import is_overridden
 from .utils import IndentedString
 
 
-CYTHON_AVAILABLE = False
-
-
 # Regular Expression for identifying a valid Python identifier name.
 _VALID_IDENTIFIER = re.compile(r'[a-zA-Z_][a-zA-Z0-9_]*')
 
@@ -174,7 +171,6 @@ class JitContext(object):
 
     """
     namespace = attr.ib(default={})  # type: Dict[str, Any]
-    use_cython = attr.ib(default=False)  # type: bool
     use_inliners = attr.ib(default=True)  # type: bool
     schema_stack = attr.ib(default=attr.Factory(set))  # type: Set[str]
     only = attr.ib(default=None)  # type: Optional[Set[str]]
@@ -631,7 +627,6 @@ def generate_marshall_method(schema, context=missing, threshold=100):
     :param schema: The Schema to generate a marshall method for.
     :param threshold: The number of calls of the same type to observe before
         specializing the marshal method for that type.
-    :param use_cython: Whether or not to attempt to use cython when Jitting.
     :return: A Callable that can be used to marshall objects for the schema
     """
     if is_overridden(schema.get_attribute, Schema.get_attribute):