--- a/toastedmarshmallow/jit.py
+++ b/toastedmarshmallow/jit.py
@@ -337,11 +337,24 @@ EXPECTED_TYPE_TO_CLASS = {
 
 def _should_skip_field(field_name, field_obj, context):
     # type: (str, fields.Field, JitContext) -> bool
-    if (getattr(field_obj, 'load_only', False) and
-            context.is_serializing):
+    load_only = getattr(field_obj, 'load_only', False)
+    dump_only = getattr(field_obj, 'dump_only', False)
+    # Marshmallow 2.x.x doesn't properly set load_only or
+    # dump_only on Method objects.  This is fixed in 3.0.0
+    # https://github.com/marshmallow-code/marshmallow/commit/1b676dd36cbb5cf040da4f5f6d43b0430684325c
+    if isinstance(field_obj, fields.Method):
+        load_only = (
+            bool(field_obj.deserialize_method_name) and
+            not bool(field_obj.serialize_method_name)
+        )
+        dump_only = (
+            bool(field_obj.serialize_method_name) and
+            not bool(field_obj.deserialize_method_name)
+        )
+
+    if (load_only and context.is_serializing):
         return True
-    if (getattr(field_obj, 'dump_only', False) and
-            not context.is_serializing):
+    if (dump_only and not context.is_serializing):
         return True
     if context.only and field_name not in context.only:
         return True