depends on commits over [here](https://github.com/lyft/marshmallow/releases/tag/2.15.1)
```bash
(venv)~/toasted-marshmallow [bump-submod*] python performance/benchmark.py 
Benchmark Result:
	Original Dump Time: 1692.94 usec/dump
	Optimized Dump Time: 243.34 usec/dump
	Original Load Time: 1092.35 usec/load
	Optimized Load Time: 238.99 usec/load
	Speed up for dump: 6.96x
	Speed up for load: 4.57x
```